правильное отображение страницы auth
