+++ src/components/ModalWindow/Modal.jsx
@@ -1,14 +1,15 @@
-import { useNavigate } from 'react-router-dom'
+// import { useNavigate } from 'react-router-dom'
 import * as S from './Modal.styles'
 
-export const Modal = () => {
-  const navigate = useNavigate()
+export const Modal = ({ childComponent }) => {
+  //   const navigate = useNavigate()
   return (
     <S.ModalDarckBG>
       <S.ModalWindow>
-        <S.Content>
-          <S.CloseBtn onClick={() => navigate(-1)}></S.CloseBtn>
-        </S.Content>
+        {/* <S.Content style={{ width: `${width}px` }}>
+          <S.CloseBtn onClick={() => navigate(-1)}></S.CloseBtn> */}
+        {childComponent}
+        {/* </S.Content> */}
       </S.ModalWindow>
     </S.ModalDarckBG>
   )
