+++ src/components/NavMenu/NavMenu.styles.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components'
+import { Button } from 'App.styles'
+
+export const Content = styled.div`
+  height: 136px;
+  width: 100%;
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  width: 100%;
+  display: flex;
+  gap: 10px;
+  justify-content: flex-start;
+  margin-top: 80px;
+`
+export const Logo = styled.div`
+  margin-right: 50px;
+`
+
+export const EnterButton = styled(Button)``
